update test hierarchy
